UrfaClient refactoring, caching support implemented
UrfaClientCollector has removed
