--- tests/Unit/UrfaClientCollectorTest.php
@@ -1,17 +1,0 @@
-<?php
-
-namespace Tests\Unit;
-
-use PHPUnit\Framework\TestCase;
-
-/**
- * @author Siomkin Alexander <siomkin.alexander@gmail.com>
- */
-class UrfaClientCollectorTest extends TestCase
-{
-    public function testSomething()
-    {
-        $this->assertTrue(true);
-    }
-
-}
